Ax `Trial`: Bug fix for error message that references nonexistent function (#2304)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2304

`BaseTrial._validate_can_attach_data` raises the following exception: f"Trial {self.index} has already been completed with data. To add more data to it (for example, for a different metric), use `Trial.update_trial_data()` or BatchTrial.update_batch_trial_data()."

This is not right because there is no `BatchTrial.update_batch_trial_data`. Also, it's an anti-pattern for a superclass to reference its subclasses like this instead of delegating to them.

This PR:
- Makes the behavior in the base class make sense for clases that have no method like "update trial data"
- Overrides that error message for `Trial` which does have such an update method

Reviewed By: bernardbeckerman, mpolson64

Differential Revision: D55439108

fbshipit-source-id: e757711972fb93374391b78af8f77d72c8abf0ed