Update some test helpers in core_stubs
Summary: I tried using `get_experiment_with_custom_runner_and_metric` and ran into some issues with model fitting due to its metrics and data not being consistent. This diff adds a couple options to the helper to make it more flexible for future use, and makes sure that the attached data corresponds to the metrics on the experiment.

Differential Revision: D55727190