Update some test helpers in core_stubs (#2319)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2319

I tried using `get_experiment_with_custom_runner_and_metric` and ran into some issues with model fitting due to its metrics and data not being consistent. This diff adds a couple options to the helper to make it more flexible for future use, and makes sure that the attached data corresponds to the metrics on the experiment.

Reviewed By: mgarrard

Differential Revision: D55727190

fbshipit-source-id: 5a96230db59842ccc9efe96b5bd169c28282f151