Tests for sequential=True as default for models in ModelBridge factory (#2324)
Summary:

Ax #1782 removed usages of sequential=True where it was silently ignored. This didn't have an effect, or at least as of today it doesn't, because sequential is set to True by default. This PR adds tests to make sure "sequential" is being passed through appropriately. The exception is cases where optimize_acqf_list is used; optimize_acqf_list is inherently sequential and doesn't accept that argument.

Differential Revision: D55772746