Tests for sequential=True as default for models in ModelBridge factory (#2324)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2324

Ax #1782 removed usages of sequential=True where it was silently ignored. This didn't have an effect, or at least as of today it doesn't, because sequential is set to True by default. This PR adds tests to make sure "sequential" is being passed through appropriately. The exception is cases where optimize_acqf_list is used; optimize_acqf_list is inherently sequential and doesn't accept that argument.

Reviewed By: saitcakmak

Differential Revision: D55772746

fbshipit-source-id: d780334ecf87bec5d853d70ba46689ee4a107add