Rename OrderedChoiceEncode => OrderedChoiceToIntegerRange (#2323)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2323

This change renames the OrderedChoiceEncode transform to one which reflects its behavior- see T182722751 for the overall task.

- Adds a new "OrderedChoiceToIntegerRange" class with the logic from the original OrderedChoiceEncode
- Updates OrderedChoiceEncode to inherit from DeprecatedTransformMixin and OrderedChoiceToIntegerRange
- Updates the registry to support the new transform.

Initially, the new classes will be decoded into the deprecated classes to maintain backwards compatibility. Once the new classes are landed, call sites will be updated to use the new class instead of the old.

Reviewed By: mpolson64

Differential Revision: D55754487

fbshipit-source-id: cb880209845f1fd45b552b74c586d2692e0e05ad