Allow bulk_fetch_trial_data to return mix of successes/failures
Summary: Currently bulk_fetch_trial_data (which we use for efficiency) will fail if any individual metric fetch fails. This can cause issues if Ax tries to fetch 2 metrics (say training loss and eval loss) and one is ready but the other is not. In the current setup Ax will report both as failures to fetch, in the new setup Ax will fetch the available one and return Err for the second

Reviewed By: sunnyshen321

Differential Revision: D55927149