Update datetime equality checks to use an absolute tolerance
Summary: Previously, this would set microseconds to 0, effectively comparing that the events happened in the same second. If the seconds differed, even when the absolute difference was negligible, the checks would fail making some tests flaky. Updating the logic to rely on an absolute difference instead, with a default tolerance of 1e-2 seconds (which is sufficient to have existing, occassionally flaky tests pass).

Differential Revision: D56077050