HSS: only check for dependents if the root parameter is present when casting parameterization (#2361)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2361

When `check_all_parameters_present=True`, we error out if any parameter is missing. When it is False, we need to tolerate any missing parameters and ignore that branch of the HSS tree -- not doing so would be ignoring the argument. The previous implementation would error out while trying to look up the value of the missing parameter if it had dependents.

Reviewed By: Balandat

Differential Revision: D56084081

fbshipit-source-id: a2cb39219fd0c14c06caa1835b9e0422713947a1