Entropy of observations metric (#2340)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2340

This commit introduces `entropy_of_observations` as a model fit metric. It quantifies the entropy of the outcomes `y_obs` using a kernel density estimator. This metric can be useful in detecting datasets in which the outcomes are clustered (implying a low entropy), rather than uniformly distributed in the outcome space (high entropy).

Reviewed By: saitcakmak

Differential Revision: D55930954

fbshipit-source-id: e2f839e6856879bae7268637966ca160aa8d6b74