Switch legacy LCE-A to qLogNEI from legacy `ei_or_nei` (#2352)
Summary:

Context for this stack: we've been meaning to deprecate ALEBO and REMBO for a long time. Tests on them break periodically and slow down development of `GenerationStrategy`, which these classes hack around, so there's substantial maintenance burden to keeping them.

Context for this diff: LCE-A was depending on ALEBO for, it seems, no reason other than convenience of where the acquisition function constructor lived. This diff aims to address.

**[RfC] But also this made me wonder: maybe better to replace use of `ei_or_nei` in LCE-A with something newer?**

Reviewed By: saitcakmak

Differential Revision: D56037477