Enable Sobol sensitivity analysis for SAAS models (#2363)
Summary:

Sobol sensitivity (D52748799) isn't available to SAAS model yet because SAAS posterior has size of b x n instead of size n (what sensitivity analysis code expects)

This diff takes the mean over b when computing posterior means of SAAS models.

We check whether the posterior is an instance of `GaussianMixturePosterior`, if yes we do `torch.mean(p.mean, dim=0)`, else we default back to the `p.mean` and `p.varaince`

Note: this doesn't work with BOTL yet but it may be worth to make it available for generic SAAS use cases first as those are the majority of the sweep. We can continue to work on making it available for BOTL/Mixarator

Differential Revision: D55999276