Enable Sobol sensitivity analysis for SAAS models (#2363)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2363

Sobol sensitivity (D52748799) isn't available to SAAS model yet because SAAS posterior has size of `b x n` instead of size `n` (what sensitivity analysis code expects), where `b` is the number of samples used in the ensemble and `n` is the `mini_batch_size`. When we do `saas_posterior.mean`, we will end up with Torch size `[b, n, 1]` and we want to `torch.cat` all posterior means by dimension 1 instead of 0 by default.

We first check whether the model is an ensemble model, if yes we cat posterior means in dim = 1, else we default back to dim = 0.

For Mixerator and BOTL, we are able to calculate unsigned sensitivity (will need to debug signed sensitivity separately) so this diff also updated `report_utils.py` to fall back to `ax_param_sens(..., signed=False)` before falling back to using lengthscales as feature importance, so that sensitivity analysis is available to BOTL and Mixerator as well.

Reviewed By: dme65

Differential Revision: D55999276

fbshipit-source-id: 8b2e99bae89b7e3248ecbd7fbe59a044f22218c3