Replace List with Sequence when typing MBM (#2394)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2394

Sequence is a more permissive type hint that works better with Pyre. For example, Pyre complains if we pass `List[MultiTaskDataset]` to a function that expects `List[SupervisedDataset]` (even though `MultiTaskDataset` is a subclass of `SupervisedDataset`) but it is ok with this if we use `Sequnce[SupervisedDataset]` as the type hint.

Reviewed By: esantorella

Differential Revision: D56498406

fbshipit-source-id: bd7c48f0bc81c1481162282d2d037b00fba7d6b9