Update GenStrategy._make_default_name to avoid fitting models
Summary: The current setup uses `node.model_spec_to_gen_from.model_key` to get the default name for a given node, which will attempt to fit models for GenNodes with multiple `model_spec`s. This diff avoids the `model_spec_to_gen_from` call for node-based GS and uses it only for step-based GS, where this is not an issue.

Differential Revision: D56732974