Use value of enum instead of casting to string (#2423)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2423

`Keys.PAIRWISE_PREFERENCE_QUERY` or `str(Keys.PAIRWISE_PREFERENCE_QUERY)` is commonly used when a string is expected, whereas I'm suspecting the intention is `Keys.PAIRWISE_PREFERENCE_QUERY.value`. This is causing issues in D56634321 which (for now) fails if any metrics in Data.df are not present on experiment when calling observations_from_dataframe.

Reviewed By: Balandat

Differential Revision: D56850035

fbshipit-source-id: d90a53ca69865dcd78410d01079757d7cadb890d