Prevent unused NaNs from marking constraints as infeasible in pareto_frontier_evaluator (#2431)
Summary:

`get_outcome_constraint_transforms` evaluates the constraints by taking the product of tensor `A` with `Y` and comparing the outcome to `rhs` (uses einsum for this). The product of `0` and `nan` evaluates to `nan`, leading to the constraint being marked infeasible, even if that `nan` is from some unused metric. By setting unused elements to `0`, this diff prevents such issues.

Reviewed By: mpolson64

Differential Revision: D56953680