handle pending observations in TimeAsFeature (#2440)
Summary:

This adds support for handling pending observations that were generated by a Modelbridge that did not use the TimeAsFeature transform and therefore does not have start_time and end_time set in the observation features.

This also adds support for untransforming observation features in Relativize that do not have a trial index. This can occur if the Modelbidge does not use TrialAsTask.

Differential Revision: D57121601