TimeAsFeature transform (#2438)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2438

This implements a transform for adding `start_time` and `duration` as features for modeling. Currently, this adds them as `RangeParameter`s (to unblock time-sensitive applications), but in the future it would be good to revise this with a better treatment of non-tunable contextual information.

`duration` appears to lead to better model fits on the synthetic example (notebook) than using `end_time`. This also works better than using the midpoint between start time and end time.

Reviewed By: bernardbeckerman, Balandat

Differential Revision: D57082939

fbshipit-source-id: d338011658918b0030b8eb3021bdec6b3276bff9