handle ObservationFeatures without trial_index in Relativize (#2441)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2441

This  adds support for untransforming observation features in Relativize that do not have a trial index. This can occur when the Modelbridge does not use TrialAsTask and there are multiple SQ observations (e.g. for different trials) and we try to untransform the predictions for arms in a GR.

Reviewed By: bernardbeckerman

Differential Revision: D57126153

fbshipit-source-id: 7e463bdaaad13850d19b7b6331fe1bd764d8e14e