model <> node 2/n: Add transition edges property to GenerationNode (#2446)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2446

As part of modifying the generation node dag we want to support multiple transition edges now (meaning one node x could transition either to node y or node z depending on what criterion are met). This property creates a dict with {'next_node_name': [tc]} where tc is the transition criterion to move the gs from the current node to the `next node`

The reason we need this, is currently and in any future i can envision, to transition to another node all criterion for that edge that are transition blocking must be met. In the next diff I modify the should_transition method. I just like bite sized diffs :)

Reviewed By: lena-kashtelyan

Differential Revision: D57127974

fbshipit-source-id: 198f8cc4a4875a36100a2d8561bdc9d9c4ac0e1f