node <> model 3/n: update should_transition_to_next_node (#2447)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2447

This diff builds on the previous diff, and uses the new property that we added for TC to decide if the gs should move forward.

If all criterion that specificy a transition edge between nodes are met, we move to that node. We progress through tc in order -- meaning when constructing a gs tc ordering is important.

Reviewed By: lena-kashtelyan

Differential Revision: D57132018

fbshipit-source-id: 5f7da7d631eab07a37988dd0196d096068234908