Add validation of transition edges to GS
Summary:
In a previous diff we decided to add validation that continue_trial_generation flag is the same for all tc that define an edge --> if these are different the gs would have conflicting information how to proceed with generation, and it would be malformed.

Also removing the transition_to optional arg is on the todo list - it requires first reworking the MaxGenerationParallelism criterion, which is why i'm doing this first.

Reviewed By: saitcakmak

Differential Revision: D57248708