Improve error clarity in TensorboardMetric (#2456)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2456

Add specific error messages to tb fetching for when data is missing.

Previously this would error out in a pandas operation which was confusing to many users. Now, we check to see whether the issue occurred because the tag was not present on the multiplexer at all (common) or because the tag was present but not data was found associated ie an empty tensor was attached (uncommon but possible).

Reviewed By: saitcakmak, eonofrey

Differential Revision: D57342319

fbshipit-source-id: 96171395844e268c1255def14878d952d582c825