Improve typing in SyntheticFunctions (#2470)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2470

A simple Pyre failure was blocking D57473132. I ended up spending way too much time trying out various options for improving typing in this file and settled on this option in the end. Updating the decorator to change the type from `Optional[T] -> T` is something I considered, but it requires the property to be typed as `Optional[T]`, which made me decide to remove the decorator altogether.

Reviewed By: Balandat

Differential Revision: D57508193

fbshipit-source-id: 4981441498dfab96c149de339af719c41e7138d4