Test that we're not referencing same parameter dict in ObservationFeatures.from_arm
Summary:
`ObservationFeatures.from_arm` is used to make the `ObservationFeatures` for the `pending_observations` dict.

Pending observations is a dict of the form `metric_name -> List[ObservationFeatures]`.  In `ModelBridge._get_transformed_gen_args`, we transform each pending `ObservationFeature` for each metric one by one. If the dict references the same object twice, we try to transform it twice, which leads to errors due to most transforms operating in-place.

~~This diff updates `ObservationFeatures.from_arm` to avoid referencing the same `parameters` dict in multiple `ObservationFeatures` objects.~~ This diff adds a test checking that `ObservationFeatures.from_arm` does not result in objects referencing the same parameters dict.
- `ObservationFeature.from_arm` is used to append the newly generated arms as pending observations in `GS._gen_multiple -> extend_pending_observations`.
- Both`get_pending_observation_features_based_on_trial_status` and `get_pending_observation_features` also use `ObservationFeature.from_arm`.

Here's a dummy example that mimics what the transform layer does to `pending_observations` (if they were to reference the same object):
{F1657963796}

Differential Revision: D57933199