Do not deduplicate against FAILED in GenerationStrategy/Node (#2488)
Summary:

We have an implicit rule that the `ABANDONED` trials failed deterministically and should not be generated again, while `FAILED` trials are unknown and may be generated again if the model chooses to do so. This rule is encoded in the utilities used to extract `pending_observations` from experiment while generating candidates, which counts `CANDIDATE` (as of D57514290), `STAGED`, `RUNNING` & `ABANDONED` trials. The `EARLY_STOPPED` and `COMPLETED` trials have data associated with them, which was a similar effect to `pending_observations` in discouraging the models from generating them again. Notably, `FAILED` trials do not fall into either category, thus they are fair game for the model to generate again, if it determines that to be the optimal next step.

This conflicts with the current deduplication against `experiment.arms_by_signature` which includes all arms on the experiment regardless of the trial status. To make the behavior consistent, this diff adds a new `arms_by_signature_for_deduplication` property on the `Experiment` that excludes arms attached to the `FAILED` trials.

Differential Revision: D57478100