Adding "analysis_batch" class to ax.analysis (#2498)
Summary:

Adding an "AnalysisBatch" class to the ax.analysis module corresponding to the XDB tables generated in this diff: D57408428

New analyses are added to the batch using the "add_analysis" method- analyses are constructed outside of the batch object because they take in different arguments.

Next steps are adding support for encoding, decoding, and saving Analysis and Analyses batch to the database.

Differential Revision: D58027893