Clean up factory functions for legacy models
Summary:
The legacy models have been raising deprecation warnings for a while, and the factory functions are generally not recommended and not supported with GS, so we can clean them up.

`get_GPEI` is left behind in this diff, since it has some usage & documentation that needs to be cleaned up along with it.

Differential Revision: D58316106