Clean up factory functions for legacy models (#2508)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2508

The legacy models have been raising deprecation warnings for a while, and the factory functions are generally not recommended and not supported with GS, so we can clean them up.

`get_GPEI` is left behind in this diff, since it has some usage & documentation that needs to be cleaned up along with it.

Reviewed By: Balandat

Differential Revision: D58316106

fbshipit-source-id: 91bf44c999a59932fa82adad9568b65445e6e27f