add option for using posterior predictive in cross-validation (#2517)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2517

see title. This change is particularly important for model selection using the NLL if we have noisy observations. Using the posterior over the true function and not the noisy observations gives quite misleading results about model calibration.

I also think that predicted vs actual plots from LOOCV are insightful when using the posterior predictive when the observations are noisy. We may want to consider adding observation_noise to `predict`, but we can do that in a follow-up.

Reviewed By: Balandat

Differential Revision: D58227612