Disallow modeling metrics not attached to the experiment (#2518)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2518

In the past it has been possible to model metrics that were not
attached to the experiment, and this has been common practice in notebook
analysis of PTS experiments where additional metrics to consider will come up
during discussion of the results for a particular batch.

This behavior was changed in D56634321, which silently drops data for metrics
that are not attached to the experiment.

I spent a while trying to figure out why my
data were being dropped.

We should either (1) allow modeling metrics that are not attached to the
experiment, or (2) not allow it, by raising an error instead of silently
dropping the data.

This diff implements (2) and raises an error that there is an issue with the metrics in the data.

This also fixes a bug in the test that was not correctly testing whether the map
metric was being included or not.

Reviewed By: dme65

Differential Revision: D57165887

fbshipit-source-id: dbb71bd0eed623ce503eb61629add267d49bfee7