Update np.Inf -> np.inf per numpy 2.0 release
Summary:
Numpy 2.0 was released yesterday ([post](https://numpy.org/news/#numpy-20-release-date-june-16)) and accordingly a couple of tests started breaking due to usage of a now-outdated API. This updates references to the now-deprecated `np.Inf` to `np.inf`, which is available in both old and new versions of numpy and so shouldn't create any issues.

Our early-stopping tutorial is still breaking due to a similar issue in TensorBoard (https://github.com/tensorflow/tensorboard/issues/6869), so not all tests are expected to pass as of now.

Differential Revision: D58672474