Ignore `sequential` argument when using `optimize_acqf_mixed`
Summary:
Passing `sequential` has been raising a deprecation warning, and stopping passing it enables  https://github.com/pytorch/botorch/pull/2390 .

Currently, the `sequential` argument is always provided by Ax and can be overridden by the user. This solution silently ignores the argument whenever the mixed optimizer is used. A nicer solution would be for Ax to only construct the `sequential` argument when it is needed and for there to be an exception when the user passes `sequential=False` and the mixed optimizer is used. If BoTorch plans to eventually enable `sequential=True` with `optimize_acqf_mixed`, then the exception should be raised by BoTorch so that Ax doesn't have to stay in sync with BoTorch's current capabilities. However, I think this code could use a thorough cleanup, so I went with the simple solution rather than add more `if` statements.

Differential Revision: D59057005