Avoid errors in telemetry due to node-based GenerationStrategy (#2554)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2554

The Ax telemetry module currently assumes that `GenerationStrategy` objects are all step-based. This diff a warning and returns a dummy record for those cases to avoid errors with node-based GSs.

The proper fix here is to fully support node-based GSs in telemetry, but that will be a bit more work it seems.

Reviewed By: saitcakmak

Differential Revision: D59193222

fbshipit-source-id: f343eddd0b78204ed5567d37864747adae8f39f7