Add back profiling (#2559)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2559

cProfile will not have the same problems we were trying to avoid in D57815360.  It is part of the standard python library.

With yappi we used to hgave a filter for the time a function ran, which is probably better than picking an arbitrary number of functions to print.  But filtering in `cProfile` would have to be done by string parsing, which feels brittle as new versions are released.

Reviewed By: esantorella

Differential Revision: D59244931

fbshipit-source-id: 117513e9e3c10d1c9faeb59aebd09e91ecc7187a