Support customizing `model_cv_kwargs` in SingleDiagnosticBestModelSelector (#2567)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2567

Context: `BestModelSelector` is used to pick the model to gen from in GNode with multiple ModelSpecs. `SingleDiagnosticBestModelSelector` is the only current implementation of it. Previously, it was limited to calling `cross_validate` with the default kwargs.

This diff adds an option to pass in `model_cv_kwargs` to `cross_validate`, which can be used to customize the options used when computing CV and corresponding diagnostics.

Reviewed By: mgarrard

Differential Revision: D59407326

fbshipit-source-id: f92af8784f8723495b29268f62e6a733d4819ace