Add an option to pass model_cv_kwargs to ModelSpec.cross_validate (#2566)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2566

This allows us to customize the kwargs passed to `cross_validate` during call time (in addition to `ModelSpec.model_cv_kwargs`, which is specified at initialization). This can be used in `SingleDiagnosticBestModelSelector` to customize how the CV diagnostics are computed.

To ensure we don't return cached results that were computed using different kwargs, we also store the last `cv_kwargs` used and re-compute CV if they changed.

Reviewed By: mgarrard

Differential Revision: D59406177

fbshipit-source-id: 577a6965aaa362d301c152ac7b94bf4fe7326541