Expand trials_as_df to support multiple GenerationNodes in a single trial
Summary:
This diff fixes a few things:
(1) Recently we expanded generation strategy to be able to support generation from multiple generation nodes for a single batch trial. However trials as df wouldn't reflect that multiple nodes or models was used for a specific trial. This diff fixes that by changes the values of the Step/Node and Model(s) column from single strs to list of strings

(2) Previously manual trials were being forgotten as they were not associated with a node, so the df was returning none. Now we return a row for manually added trials with Step/Node and Models == Manual. (Or manual would be in the list if multple nodes were used)

Differential Revision: D59738292