Add additional validation to transition criterion to not allow `transition_to` to be undefined except for MaxGenParallelism (#2579)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2579

This recently came up in a conversation with Lena/Sait when debugging an issue with transition logic (it ended up being related to the GeneratorRun node_name property being improperly set), but we should still add additional validation.

All TC should define a `transition_to` arg unless it's MaxGenParallelism. We update the logic to raise an error, and we update associated tests.

In the future, we want to reconfigure TC to never allow undefined `transition_to` so we also add a TODO

Reviewed By: lena-kashtelyan

Differential Revision: D59740485