Minor updates to AxClient.fit_model (#2580)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2580

I came across `AxClient.fit_model` while working on something else. It appeared to be a method that was split off from `get_model_predictions`, with leftover error messages that did not fit well with it as a standalone non-protected method.

This diff clears up the error messages and simplifies the definition of `AxClient.fit_model`.

Differential Revision: D59778614