Fix incorrect status quo update warning
Summary:
This is something Ben brought up as something to improve a couple months ago. Right now, with almost everything we do in any of the clients there is a warning about status quo being updated, even though it is often not being updated. It turns out this is because of the incorrect positioning of the warning in the status quo setter.

In this diff I
1) move the logger warning to be under the self._status_quo is None check
2) added a unit test modification

Differential Revision: D59894197