Raise error if enum not provided for a value we're trying to encode (#2587)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2587

We probably want to let the user know if they're trying to encode/save a value we're going to drop.  It turns out prior to this diff, a `None` for the enum that would be used to encode the value is treated differently than an actual enum that just happens to be missing the value.  So a missing enum for experiment type would not raise an error if experiment type was passed.

This should debatably only be done in test mode though.  Silent failures are generally bad though and we could cause more problems down the road by not saving data we need to.

Reviewed By: Cesar-Cardoso

Differential Revision: D59925061

fbshipit-source-id: ddea392bc67c9ddf8e966fb89982ff86465e3d1d