Add an `optimal_value` to mixed-integer problems (#2596)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2596

The next PR will require all problems to have an  `optimal_value`. See that PR for more context. This PR adds an optimal value to the mixed-integer problems, which are the only problems that currently lack one. This will be better than the true optimal value because the mixed problems have a restricted search space.

Reviewed By: saitcakmak

Differential Revision: D60144693