Rename `maximum_hypervolume` to `optimal_value` in MOO Benchmark Problems
Summary:
Context:

This will give single-objective and multi-objective benchmark problems a more consistent API, making it possible to simplify the inheritance structure in D60145193 and subsequent PRs.

Note: This breaks backwards compatibility, including for reading multi-objective benchmark problems stored as JSON. I think it's worthwhile because this will enable a lot of simplification in the future and because I expect we will want to break backward-compatibility again shortly. It would be possible to make this change BC, but I'm not sure it's worthwhile.

Note: Docstrings are updated in D60145193, so they are not updated here.

This PR:
* Makes `MultiObjectiveBenchmarkProblem.__init__` take an `optimal_value` argument, and removes its `maximum_hypervolume` argument.
* Removes the `optimal_value` property of `MultiObjectiveBenchmarkProblem`, which pointed to `maximum_hypervolume`.
* Does the same for surrogate benchmark problems.
* Updates references to `maximum_hypervolume`.

Differential Revision: D60194654