Change blob type to str
Summary:
Change the blob on AnalysisCard to always be a string rather than Any. This does not change anything for the base Anaysis or MarkdownAnalysis, but changes the dataflow on PlotlyAnalysis to use pio.to_json and pio.from_json during compute (both are very cheap).

Doing things this way allows us to decode any AnaysisCard is the same regardless of its subclass and the `card == decoded(encoded))` relatioship holds.

Reviewed By: Cesar-Cardoso

Differential Revision: D60538382