Remove status quo weight override from COPY_DB_IDS_ATTRS_TO_SKIP (#2615)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2615

Encoder and decoder both deal with `_status_quo_weight_override`.  In decoder in happens through the status quo generator run.  The problem is when a trial has a `_status_quo_weight_override`, but no status quo.  The solution in this diff is to make it impossible (unless you use protected fields directly) to have a `_status_quo_weight_override` without a `status_quo`.

## How could this be wrong?
If the user needs to store a status quo weight override on the trial for later but does not yet have a status quo.  But I don't know why they could only calculate the weight now and not later.

Reviewed By: mgarrard

Differential Revision: D60413211

fbshipit-source-id: 5bd3dbec18fa3593a1ee43119643ec70c1c4f5d6