Drop `critertion` from AutoTransitionAfterGenCriterion name
Summary:
Lena pointed out in a previous diff there is some inconsistency in the naming of transitioncriterion.

This diff creates the following pattern: if a class is a base class for other criterion (transitioncriterion, trialbasedcriterion) it keeps criterion at the end of the name since it is a bucket of criterion, and isn't directly used (except for comparisions to see if a single critierion is of that type: aka maxtrials is of trialbasedcriterion). If a criterion is a leaf of the inheritance tree, it does not have criterion at the end of the name.

Differential Revision: D60526927