Change blob type to str (#2624)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2624

Change the blob on AnalysisCard to always be a string rather than Any. This does not change anything for the base Anaysis or MarkdownAnalysis, but changes the dataflow on PlotlyAnalysis to use pio.to_json and pio.from_json during compute (both are very cheap).

Doing things this way allows us to decode any AnaysisCard is the same regardless of its subclass and the `card == decoded(encoded))` relatioship holds.

Reviewed By: Cesar-Cardoso

Differential Revision: D60538382

fbshipit-source-id: 08d4bdf7cb5e4a6c5784fc4d5ea61d70488a8bd5