Check for Nans and Infs in TensorboardMetric (#2628)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2628

Raises a ValueError in bulk_fetch_trial_data if a Nan or an Inf is found. This will get wrapped up in a MetricFetchE and handled appropriately in the Scheduler (ex. INFO if we intend to try and fetch again, WARN if coming from a tracking metric, mark trial as ABANDONED if the metric is needed for the optimization https://fburl.com/code/eq37gghi).

Differential Revision: D60670356