Drop `critertion` from AutoTransitionAfterGenCriterion name (#2625)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/2625

Lena pointed out in a previous diff there is some inconsistency in the naming of transitioncriterion.

This diff creates the following pattern: if a class is a base class for other criterion (transitioncriterion, trialbasedcriterion) it keeps criterion at the end of the name since it is a bucket of criterion, and isn't directly used (except for comparisions to see if a single critierion is of that type: aka maxtrials is of trialbasedcriterion). If a criterion is a leaf of the inheritance tree, it does not have criterion at the end of the name.

Differential Revision: D60526927

fbshipit-source-id: 40e908226f6a819c2e53f3c25fd42f87f3b84aa2